requirements.txt
# Core dependencies
playwright==1.54.0
beautifulsoup4==4.13.4
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
requests==2.32.4
pillow==11.3.0
python-multipart==0.0.20
aiofiles==24.1.0

# Development dependencies
black==25.1.0
flake8==7.3.0
pytest-cov==6.2.1
pytest-asyncio==1.1.0
mypy==1.17.0
pre-commit==4.2.0